requirements.txt
# Requirements without Version Specifiers

# nose
# nose-cov
# beautifulsoup4

python-telegram-bot
lxml
beautifulsoup4
requests

# Requirements with Version Specifiers ######`

# docopt == 0.6.1             # Version Matching. Must be version 0.6.1
# keyring >= 4.1.1            # Minimum version 4.1.1
# coverage != 3.5             # Version Exclusion. Anything except version 3.5
# Mopidy-Dirble ~= 1.1        # Compatible release. Same as >= 1.1, == 1.*

# Source: https://note.nkmk.me/en/python-pip-install-requirements/


